Revert "Merge branch 'main' of https://github.com/CodeLikeCXK/DagorEngine"
This reverts commit 416b43f4c108c0e4c425e782e341e3021eca9c9a, reversing
changes made to 982264f2a1390b6f841f9fc4845acdae57ca909c.

+++ prog/engine/drv/drv3d_vulkan/buffer.cpp
@@ -1,7 +1,5 @@
-// Copyright (C) Gaijin Games KFT.  All rights reserved.
-
 #include "buffer.h"
-#include <drv/3d/dag_commands.h>
+#include <3d/dag_drv3dCmd.h>
 #include <debug/dag_debug.h>
 #include <generic/dag_tab.h>
 #include <math/dag_TMatrix.h>
@@ -16,28 +14,36 @@
 
 #include "driver.h"
 #include "statStr.h"
-#include "device_context.h"
-#include "global_lock.h"
 
 using namespace drv3d_vulkan;
 
 GenericBufferInterface::GenericBufferInterface(uint32_t struct_size, uint32_t element_count, uint32_t flags, FormatStore format,
-  bool managed, const char *stat_name) :
+  const char *stat_name) :
   structSize(struct_size), bufSize(struct_size * element_count), bufFlags(flags), uavFormat(format)
 {
-  G_ASSERTF(bufSize, "vulkan: trying to create empty buffer <%s>", stat_name);
-  G_ASSERTF((bufFlags & (SBCF_FRAMEMEM | SBCF_USAGE_ACCELLERATION_STRUCTURE_BUILD_SCRATCH_SPACE)) !=
-              (SBCF_FRAMEMEM | SBCF_USAGE_ACCELLERATION_STRUCTURE_BUILD_SCRATCH_SPACE),
-    "vulkan: can't create RT scratch buffer as framemem");
   setResName(stat_name);
   G_ASSERTF(
     format == FormatStore(0) || ((flags & (SBCF_BIND_VERTEX | SBCF_BIND_INDEX | SBCF_MISC_STRUCTURED | SBCF_MISC_ALLOW_RAW)) == 0),
     "can't create buffer with texture format which is Structured, Vertex, Index or Raw");
 
+  Device &device = get_device();
+
+  buffer = device.createBuffer(bufSize, getMemoryClass(), getInitialDiscardCount(), BufferMemoryFlags::NONE);
+
+  device.setBufName(buffer, getResName());
+
+  if (!isDMAPathAvailable() && stagingBufferIsPermanent())
+  {
+    stagingBuffer =
+      device.createBuffer(bufSize, getPermanentStagingBufferMemoryClass(), getInitialDiscardCount(), BufferMemoryFlags::NONE);
+    stagingBuffer->setStagingDebugName(buffer);
+  }
+
   if (bufFlags & SBCF_MISC_ALLOW_RAW)
   {
     G_ASSERT(struct_size == 1 || struct_size == 4);
     structSize = 4;
+    uavFormat = FormatStore::fromCreateFlags(TEXFMT_R32UI);
   }
 
 #if DAGOR_DBGLEVEL > 0
@@ -45,69 +51,20 @@ GenericBufferInterface::GenericBufferInterface(uint32_t struct_size, uint32_t el
   {
     if (!is_good_buffer_structure_size(structSize))
     {
-      D3D_ERROR("The structure size of %u of buffer %s has a hardware unfriendly size and probably "
-                "results in degraded performance. For some platforms it might requires the shader "
-                "compiler to restructure the structure type to avoid layout violations and on other "
-                "platforms it results in wasted memory as the memory manager has to insert extra "
-                "padding to align the buffer properly.",
+      logerr("The structure size of %u of buffer %s has a hardware unfriendly size and probably "
+             "results in degraded performance. For some platforms it might requires the shader "
+             "compiler to restructure the structure type to avoid layout violations and on other "
+             "platforms it results in wasted memory as the memory manager has to insert extra "
+             "padding to align the buffer properly.",
         structSize, getResName());
     }
   }
-
-  if ((bufFlags & SBCF_ZEROMEM) && (bufFlags & SBCF_FRAMEMEM))
-  {
-    D3D_ERROR(
-      "Using SBCF_ZEROMEM and SBCF_FRAMEMEM flags for buffer %s together don't make really any sense and it prevents the driver "
-      "to apply optimizations on the buffer handling");
-  }
 #endif
 
-  // allow option-disable true framemem implementation, due to lingering amount of bugs in client usage code
-  if (!Globals::cfg.bits.allowFrameMem)
-    bufFlags &= ~SBCF_FRAMEMEM;
-  bool frameMemRequested = bufFlags & SBCF_FRAMEMEM;
-
-  if ((FormatStore(0) != uavFormat) && frameMemRequested)
-  {
-    // silently drop the flag, because we do have "one_frame" class for sampled buffers
-    // if it used in bad manner i.e. with variable discard size, we have D3D_ERROR there
-    // D3D_ERROR("vulkan: framemem for sampled buffers are not supported, dropping framemem flag for <%s>", stat_name);
-    bufFlags &= ~SBCF_FRAMEMEM;
-    frameMemRequested = false;
-  }
-
-#if DAGOR_DBGLEVEL > 0
-  // use default discard impl, it allows us to keep naming of buffer at framemem check time
-  if (Globals::cfg.bits.debugFrameMemUsage)
-    bufFlags &= ~SBCF_FRAMEMEM;
-#endif
-
-  if (!managed || (bufFlags & SBCF_FRAMEMEM))
-    return;
-
-  // pass framemem flag to object for debug purposes
-  BufferMemoryFlags memFlags = BufferMemoryFlags::NONE;
-  if (frameMemRequested)
-    memFlags = BufferMemoryFlags::FRAMEMEM;
-  else if (bufFlags & SBCF_USAGE_ACCELLERATION_STRUCTURE_BUILD_SCRATCH_SPACE)
-    memFlags = BufferMemoryFlags::DEDICATED;
-  ref = BufferRef{Buffer::create(bufSize, getMemoryClass(), 1, memFlags)};
-  afterBufferResourceAllocated();
-}
-
-void GenericBufferInterface::useExternalResource(Buffer *resource)
-{
-  ref = BufferRef{resource};
-  afterBufferResourceAllocated();
-}
-
-void GenericBufferInterface::afterBufferResourceAllocated()
-{
-  Globals::Dbg::naming.setBufName(ref.buffer, getResName());
   if (FormatStore(0) != uavFormat)
-    ref.buffer->addBufferView(uavFormat);
+    device.addBufferView(buffer, uavFormat);
 
-  if (bufFlags & SBCF_ZEROMEM)
+  if (flags & SBCF_ZEROMEM)
   {
     unsigned char *data = nullptr;
     if (lock(0, bufSize, (void **)&data, VBLOCK_WRITEONLY) && data)
@@ -120,19 +77,12 @@ void GenericBufferInterface::afterBufferResourceAllocated()
 
 GenericBufferInterface::~GenericBufferInterface()
 {
-  auto &ctx = Globals::ctx;
+  auto &ctx = get_device().getContext();
 
   if (stagingBuffer)
     ctx.destroyBuffer(stagingBuffer);
 
-  if (bufFlags & SBCF_FRAMEMEM)
-  {
-    if (ref)
-      Frontend::frameMemBuffers.release(ref.buffer);
-    return;
-  }
-
-  ctx.destroyBuffer(ref.buffer);
+  ctx.destroyBuffer(buffer);
 }
 
 int GenericBufferInterface::ressize() const { return int(bufSize); }
@@ -148,49 +98,45 @@ int GenericBufferInterface::getNumElements() const { return structSize ? bufSize
 bool GenericBufferInterface::copyTo(Sbuffer *dst)
 {
   auto dest = (GenericBufferInterface *)dst;
-  auto &context = Globals::ctx;
-  G_ASSERTF(bufSize <= ref.buffer->getBlockSize(), "the actual backing buffer is smaller (%d bytes) than the Buffer (%d bytes)",
-    ref.buffer->getBlockSize(), bufSize);
-  // reorder buffer update if we running it from external thread
-  if (Globals::lock.isAcquired())
-    context.copyBuffer(ref, dest->ref, 0, 0, bufSize);
-  else
-    context.uploadBufferOrdered(ref, dest->ref, 0, 0, bufSize);
+  auto &device = get_device();
+  auto &context = device.getContext();
+  context.copyBuffer(buffer, dest->buffer, 0, 0, bufSize);
   return true;
 }
 
 bool GenericBufferInterface::copyTo(Sbuffer *dst, uint32_t dst_offset, uint32_t src_offset, uint32_t size_bytes)
 {
   auto dest = (GenericBufferInterface *)dst;
-  auto &context = Globals::ctx;
-  // reorder buffer update if we running it from external thread
-  if (Globals::lock.isAcquired())
-    context.copyBuffer(ref, dest->ref, src_offset, dst_offset, size_bytes);
-  else
-    context.uploadBufferOrdered(ref, dest->ref, src_offset, dst_offset, size_bytes);
+  auto &device = get_device();
+  auto &context = device.getContext();
+  context.copyBuffer(buffer, dest->buffer, src_offset, dst_offset, size_bytes);
   return true;
 }
 
 uint32_t GenericBufferInterface::acquireStagingBuffer()
 {
-  TIME_PROFILE(vulkan_acquire_staging_buf);
-  // staging can be already allocated with async readback logic
-  if (!stagingBuffer)
+  if (!stagingBufferIsPermanent())
   {
-    // alloc temp with best possible memory type and smallest size only covering
-    // locked area
-    stagingBuffer = Buffer::create(lockRange.size(), getTemporaryStagingBufferMemoryClass(), 1,
-      bufferLockedForRead() ? BufferMemoryFlags::NONE : BufferMemoryFlags::TEMP);
-    stagingBuffer->setStagingDebugName(ref.buffer);
+    // staging can be already allocated with async readback logic
+    if (!stagingBuffer)
+    {
+      // alloc temp with best possible memory type and smallest size only covering
+      // locked area
+      stagingBuffer = get_device().createBuffer(lockRange.size(), getTemporaryStagingBufferMemoryClass(), 1,
+        stagingBufferIsPermanent() ? BufferMemoryFlags::NONE : BufferMemoryFlags::TEMP);
+      stagingBuffer->setStagingDebugName(buffer);
+    }
+    return 0;
   }
-  return 0;
+  else
+    return lockRange.front();
 }
 
 void GenericBufferInterface::disposeStagingBuffer()
 {
-  if (stagingBuffer)
+  if (!stagingBufferIsPermanent() && stagingBuffer)
   {
-    auto &ctx = Globals::ctx;
+    auto &ctx = get_device().getContext();
 
     ctx.destroyBuffer(stagingBuffer);
     stagingBuffer = nullptr;
@@ -201,65 +147,38 @@ void GenericBufferInterface::processDiscardFlag()
 {
   if (bufferLockDiscardRequested())
   {
-    auto &ctx = Globals::ctx;
-    // FIXME: clang on windows with dbg config, generates invalid code when .back() is used, used .stop instead
-    uint32_t dynamic_size = (bufFlags & SBCF_FRAMEMEM) > 0 ? lockRange.stop : 0;
+    auto &ctx = get_device().getContext();
 
-    ref = ctx.discardBuffer(ref, getMemoryClass(), uavFormat, bufFlags, dynamic_size);
+    buffer = ctx.discardBuffer(buffer, getMemoryClass(), uavFormat, bufFlags);
+    if (!isDMAPathAvailable() && stagingBufferIsPermanent())
+      stagingBuffer = ctx.discardBuffer(stagingBuffer, getPermanentStagingBufferMemoryClass(), FormatStore(), bufFlags);
   }
+}
+
+void GenericBufferInterface::markDataUpdate() { lastUpdateWorkId = get_device().getContext().getCurrentWorkItemId(); }
+
+bool GenericBufferInterface::updatedInCurrentWorkItem()
+{
+  return lastUpdateWorkId == get_device().getContext().getCurrentWorkItemId();
 }
 
 void GenericBufferInterface::markAsyncCopyInProgress()
 {
-  auto &ctx = Globals::ctx;
-  if (asyncCopyEvent.isRequested())
-  {
-    if (!asyncCopyEvent.isCompleted())
-    {
-      D3D_ERROR("vulkan: async readback is already requested for buffer [%s]", ref.buffer->getDebugName());
-      TIME_PROFILE(vulkan_double_readback_wait);
-      ctx.waitForIfPending(asyncCopyEvent);
-    }
-    // valid code path, we can "ignore" readback result sometimes
-    // else
-    //  D3D_ERROR("vulkan: async readback is not consumed [%s]", ref.buffer->getDebugName());
-    asyncCopyEvent.reset();
-  }
-  asyncCopyEvent.request(ctx.getCurrentWorkItemId());
+  asyncCopyInProgress = true;
   lastLockFlags = 0;
 }
 
 void GenericBufferInterface::markAsyncCopyFinished()
 {
-  if (!asyncCopyEvent.isCompleted())
-  {
-    D3D_ERROR("vulkan: async readback is not yet completed for buffer [%s]", ref.buffer->getDebugName());
-    TIME_PROFILE(vulkan_incomplete_readback_wait);
-    Globals::ctx.waitForIfPending(asyncCopyEvent);
-  }
-  asyncCopyEvent.reset();
+  // TODO: check that assumed finished copy is really finished
+  asyncCopyInProgress = false;
 }
 
-void GenericBufferInterface::blockingReadbackWait()
-{
-  TIME_PROFILE(vulkan_buffer_blocking_readback_wait);
-  Globals::ctx.wait();
-}
+void GenericBufferInterface::blockingReadbackWait() { get_device().getContext().wait(); }
 
 bool GenericBufferInterface::isDMAPathAvailable()
 {
-  if (!Globals::cfg.bits.allowDMAlockPath)
+  if (!get_device().getFeatures().test(DeviceFeaturesConfig::ALLOW_BUFFER_DMA_LOCK_PATH))
     return false;
-  return ref.buffer->hasMappedMemory();
-}
-
-BufferRef GenericBufferInterface::fillFrameMemWithDummyData()
-{
-  G_ASSERTF(bufFlags & SBCF_FRAMEMEM, "vulkan: binding empty non framemem buffer %p:%s", this, getResName());
-  // because we do like to create-bind-discard-draw, and we can't state-replace empty buffers
-  // always allocate empty block when non discarded buffer is bound
-  void *ptr;
-  lock(0, 0, &ptr, VBLOCK_WRITEONLY | VBLOCK_DISCARD);
-  unlock();
-  return ref;
+  return buffer->hasMappedMemory();
 }
